Add tower images & table sorting
Images are now displayed with tower enchantments to help easily identify
which towers they apply to. Table data can now be sorted by clicking the
columns. Refactors the data loads to be more DRY, and adds the values
directly to the anointment objects so we can manipulate them as arrays
for table sorting. Organizes images into proper directories.